dodawanie kursów, lista kursów i dodawanie trenera

+++ src/main/java/pl/sda/bootcamp/model/Course.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.Date;
 
 @AllArgsConstructor
@@ -19,7 +20,7 @@ public class Course {
     private Date startDate;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date endDate;
-    private BigDecimal price;
+    private Integer price;
     private String mode;
     private String trainer;
